fix  get user friend , some logic

+++ Social-Network-API/src/main/java/com/example/Social/Network/API/Model/ResDto/friend_res_dto/request_friend_res_dto/GetRequestFriendResDetailDto.java
@@ -22,7 +22,7 @@ public class GetRequestFriendResDetailDto {
     @JsonProperty
     String avatar;
     @JsonProperty
-    int sameFriends;
+    Long sameFriends;
 //    Time to receive requests
     @JsonProperty
     Date created;
