+++ Social-Network-API/src/main/java/com/example/Social/Network/API/Model/ResDto/friend_res_dto/user_friend_res_dto/GetUseFriendsResDetailDto.java
@@ -21,7 +21,7 @@ public class GetUseFriendsResDetailDto {
     String avatar;
 
     @JsonProperty
-    int sameFriends;
+    Long sameFriends;
     //    Time to receive requests
     @JsonProperty
     Date created;
